feat: accept prompts from stdin, including refactoring to package structure
