+++ cmd/sgpt/cli/image.go
@@ -1,19 +1,15 @@
 package cli
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"flag"
 	"fmt"
-	"image"
-	"image/png"
-	"os"
-	"path/filepath"
 	"strings"
 
-	"github.com/google/uuid"
+	"github.com/tbckr/sgpt/file"
+
+	"github.com/tbckr/sgpt/shell"
+
 	"github.com/peterbourgon/ff/v3/ffcli"
 	"github.com/sashabaranov/go-openai"
 	"github.com/tbckr/sgpt"
@@ -47,11 +43,10 @@ var imageArgs struct {
 }
 
 func runImage(ctx context.Context, args []string) error {
-	// Check, if prompt was provided via command line
-	if len(args) != 1 {
-		return ErrMissingPrompt
+	prompt, err := shell.GetPrompt(args)
+	if err != nil {
+		return err
 	}
-	prompt := args[0]
 
 	options := sgpt.ImageOptions{
 		Count: imageArgs.count,
@@ -65,7 +60,8 @@ func runImage(ctx context.Context, args []string) error {
 		responseFormat = openai.CreateImageResponseFormatURL
 	}
 
-	client, err := sgpt.CreateClient()
+	var client *openai.Client
+	client, err = sgpt.CreateClient()
 	if err != nil {
 		return err
 	}
@@ -75,6 +71,7 @@ func runImage(ctx context.Context, args []string) error {
 	if err != nil {
 		return err
 	}
+
 	return handleImageURLs(imageData)
 }
 
@@ -82,7 +79,7 @@ func handleImageURLs(images []string) error {
 	for _, data := range images {
 		if imageArgs.download {
 			// Save base64 encoded data to file
-			if err := save2File(data); err != nil {
+			if err := file.SaveEncodedImage(imageArgs.filename, data, stdout); err != nil {
 				return err
 			}
 		} else {
@@ -93,117 +90,4 @@ func handleImageURLs(images []string) error {
 		}
 	}
 	return nil
-}
-
-func save2File(imageData string) error {
-	imgBytes, err := base64.StdEncoding.DecodeString(imageData)
-	if err != nil {
-		return err
-	}
-
-	reader := bytes.NewReader(imgBytes)
-	var img image.Image
-	img, err = png.Decode(reader)
-	if err != nil {
-		return err
-	}
-
-	var filename string
-	filename, err = getFilename()
-	if err != nil {
-		return err
-	}
-
-	var file *os.File
-	file, err = os.Create(filename)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	if err = png.Encode(file, img); err != nil {
-		return err
-	}
-
-	_, err = fmt.Fprintln(stdout, filename)
-	return err
-}
-
-func getFilename() (string, error) {
-	var filename string
-	var err error
-
-	// Generate a filename, if no default filename was provided
-	if imageArgs.filename == "" {
-		filename, err = generateFilename()
-		if err != nil {
-			return "", err
-		}
-	} else {
-		filename = imageArgs.filename
-	}
-
-	// Check, if file already exists
-	var exists bool
-	baseFilename := strings.Clone(filename)
-	for {
-		exists, err = fileExists(filename)
-		if err != nil {
-			return "", err
-		}
-		// If file exists, append random string
-		if exists {
-			filename, err = makeUnique(baseFilename)
-			if err != nil {
-				return "", err
-			}
-		} else {
-			break
-		}
-	}
-	return filename, nil
-}
-
-func generateFilename() (string, error) {
-	u, err := uuid.NewRandom()
-	if err != nil {
-		return "", err
-	}
-	return u.String() + defaultImageFiletype, nil
-}
-
-func fileExists(filename string) (bool, error) {
-	// If file exists, append random string
-	if _, err := os.Stat(filename); err != nil {
-		if os.IsNotExist(err) {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func makeUnique(filename string) (string, error) {
-	// Get filename and extension
-	ext := filepath.Ext(filename)
-	name := filename[:len(filename)-len(ext)]
-
-	if ext == "" {
-		ext = defaultImageFiletype
-	}
-
-	// Generate a random byte slice
-	b := make([]byte, 6)
-	_, err := rand.Read(b)
-	if err != nil {
-		return "", err
-	}
-
-	// Encode the byte slice as a string using base64 encoding
-	randomSuffix := base64.URLEncoding.EncodeToString(b)
-
-	if name == "" {
-		return fmt.Sprintf("%s%s", randomSuffix, ext), nil
-	}
-	return fmt.Sprintf("%s_%s%s", name, randomSuffix, ext), nil
 }
